Move contractor search from HomeController to ContractorController
