+++ STAR.Web/Controllers/ContractorController.cs
@@ -1,6 +1,4 @@
 ﻿using STAR.Data;
-using STAR.Domain;
-using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
 using System.Web.Mvc;
@@ -13,17 +11,25 @@ namespace STAR.Web.Controllers {
             this.context = context as StarContext;
         }
 
-        public ActionResult Index() {
+        public ActionResult Index(string searchTerm) {
+            if (!string.IsNullOrEmpty(searchTerm)) {
+                ViewBag.SearchTerm = searchTerm;
+                //given a skill return all contractors with that skill
+                using (context) {
+                    context.Database.Log = s => System.Diagnostics.Debug.WriteLine(s);
+                    var contractors = context.Contractors
+                                  .Include(c => c.Skills)
+                                  .Where(c => c.Skills.Any(s => s.Name == searchTerm)).ToList();
+
+                    return PartialView("ContractorListPartial", contractors);
+                }
+            }
+
             return View(context.Contractors.ToList());
         }
 
         public ActionResult Details() {
             return View();
-        }
-
-        public ActionResult Search() {
-
-            return View(context.Contractors.ToList());
         }
     }
 }
